+++ bika/lims/browser/analysisrequest/view.py
@@ -146,8 +146,7 @@ class AnalysisRequestViewView(BrowserView):
 
         # If is a retracted AR, show the link to child AR and show a warn msg
         if workflow.getInfoFor(ar, 'review_state') == 'invalid':
-            childar = hasattr(ar, 'getChildAnalysisRequest') \
-                        and ar.getChildAnalysisRequest() or None
+            childar = ar.getRetest() or None
             message = _('These results have been withdrawn and are '
                         'listed here for trace-ability purposes. Please follow '
                         'the link to the retest')
@@ -158,14 +157,13 @@ class AnalysisRequestViewView(BrowserView):
             self.addMessage(message, 'warning')
         # If is an AR automatically generated due to a Retraction, show it's
         # parent AR information
-        if hasattr(ar, 'getParentAnalysisRequest') \
-            and ar.getParentAnalysisRequest():
-            par = ar.getParentAnalysisRequest()
+        invalidated = ar.getInvalidated()
+        if invalidated:
             message = _('This Analysis Request has been '
                         'generated automatically due to '
                         'the retraction of the Analysis '
                         'Request ${retracted_request_id}.',
-                        mapping={'retracted_request_id': par.getId()})
+                        mapping={'retracted_request_id': invalidated.getId()})
             self.addMessage(message, 'info')
         self.renderMessages()
         return self.template()
@@ -468,8 +466,7 @@ class AnalysisRequestViewView(BrowserView):
         workflow = getToolByName(self.context, 'portal_workflow')
         # If is a retracted AR, show the link to child AR and show a warn msg
         if workflow.getInfoFor(ar, 'review_state') == 'invalid':
-            childar = hasattr(ar, 'getChildAnalysisRequest') \
-                        and ar.getChildAnalysisRequest() or None
+            childar = ar.getRetest() or None
             anchor = childar and ("<a href='%s'>%s</a>" % (childar.absolute_url(), childar.getId())) or None
             if anchor:
                 custom['ChildAR'] = {
@@ -478,10 +475,10 @@ class AnalysisRequestViewView(BrowserView):
                 }
         # If is an AR automatically generated due to a Retraction, show it's
         # parent AR information
-        if hasattr(ar, 'getParentAnalysisRequest') \
-            and ar.getParentAnalysisRequest():
-            par = ar.getParentAnalysisRequest()
-            anchor = "<a href='%s'>%s</a>" % (par.absolute_url(), par.getId())
+        invalidated = ar.getInvalidated()
+        if invalidated:
+            anchor = "<a href='%s'>%s</a>" % (invalidated.absolute_url(),
+                                              invalidated.getId())
             custom['ParentAR'] = {
                 'title': t(_("Invalid AR retested")),
                 'value': anchor